refactor(courtboard): extract mobile modal components
